refactor: issues
